tumornet-backend/requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
Pillow==10.1.0
keras==2.15.0
numpy==1.26.1
tensorflow==2.15.0